+++ src/Symfony/Component/Scheduler/Tests/Fixtures/AttributeTask.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Scheduler\Tests\Fixtures;
+
+use Symfony\Component\Scheduler\Attribute\AsPeriodicTask;
+
+#[AsPeriodicTask(frequency: '1 hour')]
+class AttributeTask
+{
+    public array $calls = [];
+
+    public function __invoke(): void
+    {
+        $this->calls[] = __FUNCTION__;
+    }
+}